Updating to latest sputnik, fixing syntax highlighting
The latest version of sputnik split a number of rocks into separate
repos, so each of these needs to be included as a submodule as well.
Also, changes to sputnik.ws are not reflected in the running site, which
uses sandbox/sputnik.cgi. This was the issue with syntax highlighting,
and this has now been resolved.